Convert API to double

+++ jsplat-app/src/main/java/de/javagl/jsplat/app/DataSet.java
@@ -98,7 +98,7 @@ class DataSet
     void setTransform(Transform transform)
     {
         this.currentTransform = transform;
-        float[] matrix = Transforms.toMatrix(transform);
+        double[] matrix = Transforms.toMatrix(transform);
         int shDimensions = Splats.dimensionsForDegree(shDegree);
         Consumer<MutableSplat> t =
             SplatTransforms.createTransform(matrix, shDimensions);
